+++ security-command-center/snippets/src/main/java/vtwo/findings/GroupFindingsWithFilter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v_two.findings;
+package vtwo.findings;
 
 // [START securitycenter_group_filtered_findings_v2]
