+++ security-command-center/snippets/src/main/java/vtwo/findings/ListAllFindings.java
@@ -14,20 +14,19 @@
  * limitations under the License.
  */
 
-package v_two.findings;
+package vtwo.findings;
 
-// [START securitycenter_list_filtered_findings_v2]
+// [START securitycenter_list_all_findings_v2]
 
 import com.google.cloud.securitycenter.v2.ListFindingsRequest;
 import com.google.cloud.securitycenter.v2.ListFindingsResponse.ListFindingsResult;
 import com.google.cloud.securitycenter.v2.SecurityCenterClient;
 import java.io.IOException;
 
-public class ListFindingsWithFilter {
+public class ListAllFindings {
 
   public static void main(String[] args) throws IOException {
-    // TODO: Replace the variables within {}
-    // organizationId: Google Cloud Organization id.
+    // organizationId: The source to list all findings for.
     // You can also use project/ folder as the parent resource.
     String organizationId = "google-cloud-organization-id";
 
@@ -37,31 +36,22 @@ public class ListFindingsWithFilter {
     // The source id to scope the findings.
     String sourceId = "source-id";
 
-    listFilteredFindings(organizationId, sourceId, location);
+    listAllFindings(organizationId, sourceId, location);
   }
 
-  // List filtered findings under a source.
-  public static void listFilteredFindings(String organizationId, String sourceId, String location)
+  // List all findings under a given parent resource.
+  public static void listAllFindings(String organizationId, String sourceId, String location)
       throws IOException {
     // Initialize client that will be used to send requests. This client only needs to be created
     // once, and can be reused for multiple requests.
     try (SecurityCenterClient client = SecurityCenterClient.create()) {
-
-      // Use any one of the following formats:
-      //  * organizations/{organization_id}/sources/{source_id}/locations/{location}
-      //  * folders/{folder_id}/sources/{source_id}/locations/{location}
-      //  * projects/{project_id}/sources/{source_id}/locations/{location}
-      String parent = String.format("organizations/%s/sources/%s/locations/%s", organizationId,
-          sourceId,
-          location);
-
-      // Listing all findings of category "MEDIUM_RISK_ONE".
-      String filter = "category=\"MEDIUM_RISK_ONE\"";
-
       ListFindingsRequest request =
           ListFindingsRequest.newBuilder()
-              .setParent(parent)
-              .setFilter(filter)
+              // To list findings across all sources, use "-".
+              .setParent(
+                  String.format("organizations/%s/sources/%s/locations/%s", organizationId,
+                      sourceId,
+                      location))
               .build();
 
       for (ListFindingsResult result : client.listFindings(request).iterateAll()) {
@@ -71,4 +61,4 @@ public class ListFindingsWithFilter {
     }
   }
 }
-// [END securitycenter_list_filtered_findings_v2]
+// [END securitycenter_list_all_findings_v2]
