+++ security-command-center/snippets/src/main/java/vtwo/muteconfig/DeleteMuteRule.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v_two.muteconfig;
+package vtwo.muteconfig;
 
 // [START securitycenter_delete_mute_config_v2]
