+++ security-command-center/snippets/src/main/java/vtwo/muteconfig/ListMuteRules.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v_two.muteconfig;
+package vtwo.muteconfig;
 
 // [START securitycenter_list_mute_configs_v2]
