+++ security-command-center/snippets/src/test/java/vtwo/Util.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v_two;
+package vtwo;
 
 import com.google.cloud.securitycenter.v2.CreateFindingRequest;
 import com.google.cloud.securitycenter.v2.CreateSourceRequest;
